feat: 'OBSERVE' command pipeline parser and observe events emitter
This is a partial prototype implementation. It doesn't work yet, but shows the direction.

Note that certain designs may not be representative of what will land in the final proposal.
For example: pipeline parser likely will have to be hand-coded, as bison+flex might bring too much complexity.
Currently defined OBSERVE commands were created just to spawn the placeholders for the future implementation
of the OBSERVE commands.